reorganizo archivos en 3 subcomponentes Map, CustomMap y MapView
