+++ frontend/src/components/composed/Map/CustomMap/CustomMapIcons/CustomMapIcons.js
