+++ frontend/src/components/composed/Map/CustomMap/IconSelector/CustomIcon/CustomIcon.js
