+++ frontend/src/components/composed/Map/CustomMap/IconSelector/IconSelector.js
