+++ frontend/src/components/composed/Map/CustomMap/IconSelector/IconsSelector.styles.js
