+++ frontend/src/components/composed/Map/MapView/MapMarkers/MapIcon/MapIcon.js
